fixed runtime class loading, database tabels and much more, finally running one game :)
