+++ examples/simple/config.toml
@@ -1,0 +1,5 @@
+[CompetitionServer]
+    port = 65335
+    log = "DEBUG"
+    game_path = "examples\\simple\\game.py"
+    game_class = "ExampleGame"
